+++ ff_data_server/src/oanda_api/get/accounts/account_details.rs
@@ -1,9 +1,57 @@
-use serde::{Deserialize};
-use serde_json::Value;
-use rust_decimal::Decimal;
+use ff_standard_lib::messages::data_server_messaging::FundForgeError;
+use serde_derive::Deserialize;
 use chrono::{DateTime, Utc};
-use crate::oanda_api::models::position::OandaPosition;
+use rust_decimal::Decimal;
+use serde_json::Value;
+use crate::oanda_api::api_client::OandaClient;
+use crate::oanda_api::get::positions::OandaPosition;
 use crate::oanda_api::models::trade::TradeSummary;
+
+pub(crate) async fn get_oanda_account_details(oanda_client: &OandaClient, account_id: &str) -> Result<OandaAccount, FundForgeError> {
+    let request_uri = format!("/accounts/{}", account_id);
+
+    let response = match oanda_client.send_rest_request(&request_uri).await {
+        Ok(response) => response,
+        Err(e) => {
+            return Err(FundForgeError::ServerErrorDebug(
+                format!("Failed to get_requests account summary from server: {:?}", e)
+            ));
+        }
+    };
+
+    if !response.status().is_success() {
+        return Err(FundForgeError::ServerErrorDebug(
+            format!("Server returned error status: {}", response.status())
+        ));
+    }
+
+    let content = match response.text().await {
+        Ok(content) => content,
+        Err(e) => {
+            return Err(FundForgeError::ServerErrorDebug(
+                format!("Failed to read response content: {:?}", e)
+            ));
+        }
+    };
+
+    let json: serde_json::Value = match serde_json::from_str(&content) {
+        Ok(json) => json,
+        Err(e) => {
+            return Err(FundForgeError::ServerErrorDebug(
+                format!("Failed to parse JSON response: {:?}", e)
+            ));
+        }
+    };
+
+    //eprintln!("Account JSON: {}", json["account"]);
+
+    match serde_json::from_value(json["account"].clone()) {
+        Ok(summary) => Ok(summary),
+        Err(e) => Err(FundForgeError::ServerErrorDebug(
+            format!("Failed to deserialize account summary: {:?}", e)
+        ))
+    }
+}
 
 #[derive(Debug, Deserialize)]
 #[allow(dead_code)]
@@ -69,54 +117,4 @@ pub(crate) struct OandaAccount {
     pub positions: Vec<OandaPosition>,
     #[serde(default)]
     pub orders: Vec<Value>,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct AccountChangesResponse {
-    pub(crate) changes: AccountChanges,
-    #[serde(rename = "lastTransactionID")]
-    pub(crate) last_transaction_id: String,
-    pub(crate) state: AccountState,
-}
-
-#[derive(Debug, Deserialize)]
-#[allow(dead_code)]
-pub struct AccountChanges {
-    #[serde(rename = "ordersCancelled")]
-    orders_cancelled: Vec<Value>,
-    #[serde(rename = "ordersCreated")]
-    orders_created: Vec<Value>,
-    #[serde(rename = "ordersFilled")]
-    orders_filled: Vec<Value>,
-    #[serde(rename = "ordersTriggered")]
-    orders_triggered: Vec<Value>,
-    pub(crate) positions: Vec<OandaPosition>,
-    #[serde(rename = "tradesClosed")]
-    trades_closed: Vec<Value>,
-    #[serde(rename = "tradesOpened")]
-    trades_opened: Vec<Value>,
-    #[serde(rename = "tradesReduced")]
-    trades_reduced: Vec<Value>,
-    transactions: Vec<Value>,
-}
-
-#[derive(Debug, Deserialize)]
-#[allow(dead_code)]
-pub struct AccountState {
-    #[serde(rename = "NAV")]
-    nav: Decimal,
-    #[serde(rename = "marginAvailable")]
-    pub(crate) margin_available: Decimal,
-    #[serde(rename = "marginUsed")]
-    pub(crate) margin_used: Decimal,
-    #[serde(rename = "positionValue")]
-    position_value: Decimal,
-    #[serde(rename = "marginCloseoutPercent")]
-    margin_closeout_percent: Decimal,
-    #[serde(rename = "marginCloseoutUnrealizedPL")]
-    margin_closeout_unrealized_pl: Decimal,
-    #[serde(rename = "unrealizedPL")]
-    pub(crate) unrealized_pl: Decimal,
-    #[serde(rename = "withdrawalLimit")]
-    withdrawal_limit: Decimal,
 }
