+++ ff_data_server/src/oanda_api/handlers/mod.rs
@@ -1,2 +1,3 @@
 pub mod account_updates;
 pub mod stream;
+pub mod quotebar_streams;
