+++ ff_data_server/src/oanda_api/mod.rs
@@ -2,10 +2,9 @@ pub mod api_client;
 pub mod base_data_converters;
 pub mod download;
 pub mod get;
-pub mod instruments;
 pub mod support_and_conversions;
 pub mod settings;
 pub mod vendor_api_response;
 pub mod broker_api_response;
-pub mod models;
 pub mod handlers;
+pub mod models;
