+++ ff_data_server/src/oanda_api/models/mod.rs
@@ -1,10 +1,9 @@
 pub mod account;
 pub mod order;
 pub mod transaction;
-pub mod instruments;
-pub mod position;
 pub mod pricing;
 pub mod pricing_common;
 pub mod primitives;
 pub mod trade;
 pub mod transaction_related;
+pub mod instruments;
